Create swap interface
