+++ pkg/bee/cluster.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ethersphere/beekeeper/pkg/k8s"
 	k8sBee "github.com/ethersphere/beekeeper/pkg/k8s/bee"
 	"github.com/ethersphere/beekeeper/pkg/k8s/notset"
+	"github.com/ethersphere/beekeeper/pkg/swap"
 )
 
 // Cluster represents cluster of Bee nodes
@@ -24,11 +25,11 @@ type Cluster struct {
 	debugAPIInsecureTLS bool
 	debugAPIScheme      string
 	k8s                 *k8s.Client
-	// swap                *swap.Client
-	labels           map[string]string
-	namespace        string
-	disableNamespace bool                  // do not use namespace for node hostnames
-	nodeGroups       map[string]*NodeGroup // set when groups are added to the cluster
+	swap                swap.Client
+	labels              map[string]string
+	namespace           string
+	disableNamespace    bool                  // do not use namespace for node hostnames
+	nodeGroups          map[string]*NodeGroup // set when groups are added to the cluster
 }
 
 // ClusterOptions represents Bee cluster options
@@ -41,10 +42,10 @@ type ClusterOptions struct {
 	DebugAPIInsecureTLS bool
 	DebugAPIScheme      string
 	K8SClient           *k8s.Client
-	// Swap                *swap.Client
-	Labels           map[string]string
-	Namespace        string
-	DisableNamespace bool
+	SwapClient          swap.Client
+	Labels              map[string]string
+	Namespace           string
+	DisableNamespace    bool
 }
 
 // NewCluster returns new cluster
@@ -59,10 +60,10 @@ func NewCluster(name string, o ClusterOptions) *Cluster {
 		debugAPIInsecureTLS: o.DebugAPIInsecureTLS,
 		debugAPIScheme:      o.DebugAPIScheme,
 		k8s:                 o.K8SClient,
-		// swap:                o.Swap,
-		labels:           o.Labels,
-		namespace:        o.Namespace,
-		disableNamespace: o.DisableNamespace,
+		swap:                o.SwapClient,
+		labels:              o.Labels,
+		namespace:           o.Namespace,
+		disableNamespace:    o.DisableNamespace,
 
 		nodeGroups: make(map[string]*NodeGroup),
 	}
